updated affirmations
