+++ functions/automations/affirmations/index.js
