+++ functions/automations/dailyCC/constants.js
