+++ functions/automations/dailyCC/dailyCC.js
