+++ functions/automations/dailyCC/index.js
